% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDfEmpty.R
\name{checkDfEmpty}
\alias{checkDfEmpty}
\title{Check if a dataframe is empty or not}
\usage{
checkDfEmpty(df, is_empty)
}
\arguments{
\item{df}{A dataframe to check.}

\item{is_empty}{Logical value. Are you expecting df to be empty?}
}
\value{
Returns \"Passed\" if the dataframe meets the condition.
}
\description{
This function stops with an error message if the dataframe does not meet the \code{is_empty} condition.
}
\examples{
\dontrun{
df <- data.frame()
checkNaDf(df, FALSE) # This will throw an error because df is empty
checkNaDf(df, TRUE)  # This will return "Pass" because df is empty
}
}
