% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNaDf.R
\name{checkNaDf}
\alias{checkNaDf}
\title{Check for NA values in a specified column of a dataframe}
\usage{
checkNaDf(df, column_name)
}
\arguments{
\item{df}{A dataframe in which to check for NA values.}

\item{column_name}{A string specifying the name of the column to check.}
}
\value{
This function does not return a value. It either throws an error or displays a message.
}
\description{
This function checks if a specified column in a dataframe contains any NA values.
If the column does not exist in the dataframe, an error is thrown.
If the column contains any NA values, an error is thrown.
If the column does not contain any NA values, a message is displayed.
}
\examples{
\dontrun{
df <- data.frame(a = c(1, 2, NA), b = c(4, 5, 6))
checkNaDf(df, "a") # This will throw an error because column 'a' contains NA values
checkNaDf(df, "b") # This will display a message because column 'b' does not contain any NA values
}
}
