% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDimIds.R
\name{genDimIds}
\alias{genDimIds}
\title{Generate dimension IDs}
\usage{
genDimIds(dimSize, dimName = NULL)
}
\arguments{
\item{dimSize}{A vector of dimension sizes}

\item{dimName}{A vecotr of dimension names}
}
\value{
A dataframe containing dimension IDs
}
\description{
Generate dimension IDs
}
\examples{
genDimIds(dimSize =  c(2,3,4), dimName = c("site_id", "period_id", "worker_id"))
}
