% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logging}
\alias{logging}
\title{Logging class}
\description{
Logging class

Logging class
}
\examples{
myLog = logging$new("log.txt")

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fileName}}{is a string}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-logging-new}{\code{logging$new()}}
\item \href{#method-logging-write}{\code{logging$write()}}
\item \href{#method-logging-createFile}{\code{logging$createFile()}}
\item \href{#method-logging-clone}{\code{logging$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logging-new"></a>}}
\if{latex}{\out{\hypertarget{method-logging-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logging$new(fileName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileName}}{The name of the logging file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An logging object
Write a message
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logging-write"></a>}}
\if{latex}{\out{\hypertarget{method-logging-write}{}}}
\subsection{Method \code{write()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logging$write(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{The message to be printed
Create a logging file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logging-createFile"></a>}}
\if{latex}{\out{\hypertarget{method-logging-createFile}{}}}
\subsection{Method \code{createFile()}}{
create a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logging$createFile()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logging-clone"></a>}}
\if{latex}{\out{\hypertarget{method-logging-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logging$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
