% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logSummary.R
\name{logSummary}
\alias{logSummary}
\title{Print model summary to the log file}
\usage{
logSummary(mdl)
}
\arguments{
\item{mdl}{an estiamted regression model}
}
\description{
Print model summary to the log file
}
